updated project structure to add back in the ui as ForestKitView, updated readme
