+++ Project.swift
@@ -22,5 +22,8 @@ import ProjectDescriptionHelpers
 // Creates our project using a helper function defined in ProjectDescriptionHelpers
 let project = Project.app(name: "Forest",
                           platform: .iOS,
-                          additionalTargets: ["ForestKit", "ForestKitView"],
+                          additionalTargets: [
+                              Project.AdditonalTarget(name: "ForestKit"),
+                              Project.AdditonalTarget(name: "ForestKitView", dependencies: ["ForestKit"])
+                          ],
                           additionalFiles: ["README.md", "LICENSE", "Package.swift"])
