+++ README.md
@@ -2,7 +2,7 @@
 
 iOS logging framework inspired by [Timber](https://github.com/JakeWharton/timber) and [square/logcat](https://github.com/square/logcat).
 
-## Useage
+## Useage ForestKit
 
 1. Install the library into your project.
     * Swift Package Manager:  `.package(url: "https://github.com/ameriod/ForestKit.git", from: "1.0.0")`
@@ -29,6 +29,10 @@ To disable logging just do not plant any trees. To remove a tree at run time cal
 
 See the [sample app](Targets/Forest/Sources) for more usage.
 
+## Usage ForestKitView
+
+The `ForestKitView.LogViewer` provides a SwiftUI view to view the logs producted by the [ForestKit.FileOutputTree](Targets/ForestKit/Sources/FileOutputTree.swift). Simplly add it the your debug menu of the app and you can search, filter,  sort logs and copy the log messages.
+
 ## Project setup
 
 The project is generated by [Tuist](https://github.com/tuist/tuist).
@@ -46,7 +50,7 @@ If there are any new files added when you pull down from git you will need to re
 ## TODOs
 
 * Add a `URLProtocol` to log network requests.
-* Include the `URLProtocol` logs in the UI
+* Include the `URLProtocol` logs in the `ForestKitView.LogViewer`
 * Add Carthage support.
 * Add Cocopods support.
 * Add CI/CD to build the project.
