+++ Targets/ForestKitView/Sources/FileOutputTreeView.swift
@@ -2,11 +2,12 @@ import Combine
 import ForestKit
 import SwiftUI
 
-public extension ForestKit {
+public extension ForestKitView {
 
-    struct FileOutputTreeView: View {
+    /// The UI to view logs producted by the `ForestKit.FileOutputTree`.
+    struct LogViewer: View {
 
-        @ObservedObject private var viewModel = ViewModel()
+        @ObservedObject private var viewModel = FileOutputTreeViewModel()
         @State private var showFilters = false
 
         public init() { }
@@ -41,7 +42,7 @@ public extension ForestKit {
 
 private struct LogView: View {
 
-    let data: FileLogData
+    let data: LogData
 
     var body: some View {
         VStack(alignment: .leading) {
@@ -58,7 +59,7 @@ private struct LogView: View {
 
 private struct LogFilterView: View {
 
-    @ObservedObject var viewModel: ViewModel
+    @ObservedObject var viewModel: FileOutputTreeViewModel
     @Binding var showFilters: Bool
 
     var body: some View {
@@ -81,7 +82,7 @@ private struct LogFilterView: View {
                         }
                 }
             }
-            Section(header: Text("Date")) {
+            Section(header: Text("Sort")) {
                 ForEach(SortOrder.allCases, id: \.self) { sort in
                     Text(sort.rawValue)
                         .foregroundColor(selectedColor(viewModel.isSelected(for: sort)))
@@ -104,16 +105,6 @@ private struct LogFilterView: View {
     }
 }
 
-private enum ErrorFilter: String, CaseIterable {
-    case yes = "Errors"
-    case no = "No Errors"
-}
-
-private enum SortOrder: String, CaseIterable {
-    case decending = "Decending Date"
-    case acending = "Accending Date"
-}
-
 extension ForestKit.Priority {
 
     var color: Color {
@@ -122,141 +113,6 @@ extension ForestKit.Priority {
             return .gray
         case .error, .fault, .wtf:
             return .red
-        }
-    }
-}
-
-private class ViewModel: ObservableObject {
-
-    private let tree: ForestKit.FileOutputTree
-    private let jsonDecoder: JSONDecoder
-    private var logsFromDisk = [FileLogData]()
-
-    @Published var logs = [FileLogData]()
-    @Published var search = ""
-    @Published var selectedPriorities = Set(ForestKit.Priority.allCases)
-    @Published var selectedError: ErrorFilter?
-    @Published var selectedSort: SortOrder = .decending
-    private var disposables = Set<AnyCancellable>()
-
-    init() {
-        if let tree = ForestKit.instance.forest().first(where: { $0 is ForestKit.FileOutputTree }) as? ForestKit.FileOutputTree {
-            self.tree = tree
-        } else {
-            print("\(ForestKit.Priority.error) - No \(ForestKit.FileOutputTree.self) is planted")
-            // Create a dummy tree since the logging library should not crash the app.
-            tree = ForestKit.FileOutputTree()
-        }
-        jsonDecoder = JSONDecoder()
-        jsonDecoder.dateDecodingStrategy = .iso8601
-
-        Publishers.CombineLatest4($search, $selectedPriorities, $selectedError, $selectedSort)
-            .map { [unowned self] query, priorities, error, sort in
-                self.logsFromDisk.search(with: query)
-                    .filter(with: priorities)
-                    .filter(with: error)
-                    .sorted(by: sort)
-            }
-            .sink { searchedLogs in
-                self.logs = searchedLogs
-            }
-            .store(in: &disposables)
-    }
-
-    func isSelected(for priority: ForestKit.Priority) -> Bool {
-        selectedPriorities.contains(priority)
-    }
-
-    func prioritySelected(with priority: ForestKit.Priority) {
-        if isSelected(for: priority) {
-            selectedPriorities.remove(priority)
-        } else {
-            selectedPriorities.insert(priority)
-        }
-    }
-
-    func isSelected(for error: ErrorFilter) -> Bool {
-        error == selectedError
-    }
-
-    func errorFilterSelected(with error: ErrorFilter) {
-        if error == selectedError {
-            selectedError = nil
-        } else {
-            selectedError = error
-        }
-    }
-
-    func isSelected(for sort: SortOrder) -> Bool {
-        sort == selectedSort
-    }
-
-    func selectedSort(with sort: SortOrder) {
-        selectedSort = sort
-    }
-
-    func load() {
-        DispatchQueue.main.async {
-            self.loadSync()
-        }
-    }
-
-    private func loadSync() {
-        do {
-            let log = try String(contentsOf: tree.fileURL, encoding: .utf8)
-            // Seperate on the end of the JSON object + the new line, it should be unique.
-            logsFromDisk = try log.components(separatedBy: "}\n")
-                .filter { !$0.isEmpty }
-                .compactMap {
-                    // Need add back in the `}` since the split removed it
-                    guard let data = "\($0)}".data(using: .utf8) else { return nil }
-                    return try jsonDecoder.decode(FileLogData.self, from: data)
-                }
-            logs = logsFromDisk
-        } catch {
-            print("\(ForestKit.Priority.error) - Getting log file for: \(tree.fileURL)")
-        }
-    }
-}
-
-private extension Array where Element == FileLogData {
-    func search(with query: String) -> [FileLogData] {
-        if query.isEmpty {
-            return self
-        } else {
-            return filter { log in
-                log.message.localizedCaseInsensitiveContains(query) ||
-                    log.error?.localizedCaseInsensitiveContains(query) == true ||
-                    log.priority.rawValue.localizedCaseInsensitiveContains(query)
-            }
-        }
-    }
-
-    func filter(with priorities: Set<ForestKit.Priority>) -> [FileLogData] {
-        filter {
-            priorities.contains($0.priority)
-        }
-    }
-
-    func filter(with error: ErrorFilter?) -> [FileLogData] {
-        filter {
-            switch error {
-            case .no:
-                return $0.error == nil
-            case .yes:
-                return $0.error != nil
-            default:
-                return true
-            }
-        }
-    }
-
-    func sorted(by sort: SortOrder) -> [FileLogData] {
-        switch sort {
-        case .acending:
-            return sorted(by: { $0.date.compare($1.date) == .orderedAscending })
-        case .decending:
-            return sorted(by: { $0.date.compare($1.date) == .orderedDescending })
         }
     }
 }
