+++ Targets/ForestKitView/Sources/ForestKitView.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+/// Name space for the view components of `ForestKit`
+public enum ForestKitView { }
