+++ Tuist/ProjectDescriptionHelpers/Project+Templates.swift
@@ -6,12 +6,23 @@ import ProjectDescription
 /// See https://docs.tuist.io/guides/helpers/
 
 extension Project {
+
+    public struct AdditonalTarget {
+        var name: String
+        var dependencies: [String]
+
+        public init(name: String, dependencies: [String] = []) {
+            self.name = name
+            self.dependencies = dependencies
+        }
+    }
+
     /// Helper function to create the Project for this ExampleApp
-    public static func app(name: String, platform: Platform, additionalTargets: [String], additionalFiles: [FileElement]) -> Project {
+    public static func app(name: String, platform: Platform, additionalTargets: [AdditonalTarget], additionalFiles: [FileElement]) -> Project {
         var targets = makeAppTargets(name: name,
                                      platform: platform,
-                                     dependencies: additionalTargets.map { TargetDependency.target(name: $0) })
-        targets += additionalTargets.flatMap { makeFrameworkTargets(name: $0, platform: platform) }
+                                     dependencies: additionalTargets.map { TargetDependency.target(name: $0.name) })
+        targets += additionalTargets.flatMap { makeFrameworkTargets(name: $0.name, dependencies: $0.dependencies, platform: platform) }
         return Project(name: name,
                        organizationName: "github.com/ameriod",
                        targets: targets,
@@ -21,7 +32,7 @@ extension Project {
     // MARK: - Private
 
     /// Helper function to create a framework target and an associated unit test target
-    private static func makeFrameworkTargets(name: String, platform: Platform) -> [Target] {
+    private static func makeFrameworkTargets(name: String, dependencies: [String], platform: Platform) -> [Target] {
         let sources = Target(name: name,
                              platform: platform,
                              product: .framework,
@@ -29,7 +40,7 @@ extension Project {
                              infoPlist: .default,
                              sources: ["Targets/\(name)/Sources/**"],
                              resources: [],
-                             dependencies: [])
+                             dependencies: dependencies.map { .target(name: $0) })
         let tests = Target(name: "\(name)Tests",
                            platform: platform,
                            product: .unitTests,
